Update README.md
The Flask Portfolio Application is a dynamic web platform designed to showcase personal achievements, professional experiences, and project work. Built using the Flask web framework, this application serves as a comprehensive portfolio that can be easily updated by modifying JSON files and HTML templates. It provides a user-friendly interface to present a timeline of events, a reading list, and various projects and experiences. The application supports filtering and sorting of projects and experiences, and includes custom error handling for a seamless user experience.